Better default arguments (fixes #289).
This patch allows default arguments to be a full expression instead of
requiring them to be a literal. This means storing them in checkpoints
and adding a function to reverse the checkpoints.

A side-effect of this is that it's reasonable to have a default value
for optional arguments. So emitter and parser's token handling were
both updated to support optargs + varargs on the same function.

This does not change how backend functions receive either kind of
argument. This is purely about getting parser+emitter to understand
that the two can now mix.