Add `%ld` to msgbuf for int64_t values. (#325)
The msgbuf used to use `int` for `%d`. It still does, but now `%ld`
exists to allow values that are explicitly 64-bits. Since Lily is
usually built on a 64-bit system, no problem was noticed.

However, when compiling to 32-bits, `int` is no longer the `int64_t`
that `Integer` uses. So if a 64-bit int is followed by another int
value, the second value won't be correct since only 32 bits are taken.

The out-of-bounds error message for `List.insert` and `List.delete_at`
does exactly that.

As a precaution, this patch also adjusts the vm's internals to now use
`%ld` when raising boundary errors.