vm: Build operations must do the move last. (fixes #329).
A while back, commit 9997c7e removed call states and merged together
variant handling with the rest of calls. Since that commit, the binary
tree benchmark has been broken.

The reason is that calls like to check if one of their inputs has a
register of the outgoing type. So variants were now doing that as
well.

The problem would show itself when writing code like
`Branch(1, Leaf(1), Leaf(1))`. The emitter would notice that the first
`Leaf` has the same enum type, and use that as the result.

Prior to this, build operations would do their move first. So the
above code would instead create a `Branch` wherein the left side was
not a `Leaf`, but the `Branch` itself!

For safety, I've adjusted all build ops instead of just variants.