Remove default scopes for classes (fixes #308).
This adds 'public' as a new keyword to the language. Now, when a
class method or var is declared, a user must choose if item is public,
protected, or private.

Before this, having no keyword meant that the item in question would
be public. At first, this seemed a natural choice for concise code.
In writing parsekit, I noticed several times I would accidentally
leave a method public (since that's the default) when I didn't
actually want a method to be public.

I thought for a time of making the default private. Eventually I
decided to do away with defaults at all. My reasons are:

* No need to remember what the default is.

* The programmer must now think over what scope is proper, instead of
  (accidentally) omitting it.

* Keyword usage can be easily sifted through using grep.

* Consistency with other languages that don't have a default scope.