Fix templates not honoring multiple tags.
The underlying problem is that end tag wasn't calling up the next
token. So after the first tag, everything was getting ignored. To make
sure this didn't happen, I decided to write a new kind of test
(render this and it needs to give an error).

Which is when I noticed that the traceback included two __main__
calls. Turns out that parser preps __main__ on every pass. So this
patch includes fixing that since the test checks traceback.