Parser: Don't allow else in a complete enum match.
I discovered this while working on the octeon debug branch. The old
test suite had a test for this that was failing. Apparently between
the rewrite of enum handling and the rewrite of the test suite, the
test was lost. I've put in a new test to make sure this bug doesn't
come back.