Parser: Remove writing a qualifier for non-__main__/builtin errors.
Also found this while working on the octeon debug branch. This time,
the test was failing because at some point I removed the dot from the
namespace qualifier.

I've been considering removing this for some time now. I think that
if someone is raising an error, that the error should have some sort
of indication of what it's for. That the error should not rely solely
on the namespace it's in for context. So, for example, a json package
might export JsonError.

Part of the rationale is that such behavior makes those packages more
friendly toward explicit same-scope include of all of their contents.
That's not in yet, but I plan to have it done in the near future.

I've also been hunting for an excuse to get rid of this, and I'm too
lazy to fix the test.