ts: Use `?` instead of NULL.
A long time ago, `Function` didn't do `Unit` returning like it does
now. The return type of a call with no output was denoted as `NULL`.
The NULL check in check_raw is one example of a now-impossible test
that was never erased when `Unit` was added.

Using `?` as a placeholder also simplifies some of the ts checking
code.