Do away with 'default to Dynamic' behavior (fixes #324).
This behavior is a leftover from a very long time ago when `Dynamic`
was instead Any. Back in those not-great days, Any was a type that
other types would be implicitly boxed upward into. So, for example,
code such as 'var v = [1, "2"]' would become `List[Dynamic]`, and then
allow push with any kind of a value.

With `Dynamic` being opt-in and not magical, defaulting to using it
makes no sense. The emitter now requires that the right side of an
assignment is always a complete type.