Parser: Remove code that is now impossible.
It used to be that creating a class method caused it to immediately be
added to vars. It would later get pulled off of the var chain and put
into the class method chain properly.

A short time ago, that was changed so that method vars were
immediately placed. Once that happened, these two sections became
impossible to execute.