Add a massive pile of tests.
There are still parts of the interpeter not covered, but not nearly as
much as before. I've decided to merge this incomplete set of test so
that each of the issues involving tests can be a commit by themselves.

Several bugfix commits pushed lately were the result of catching
problems while developing these tests.

I'm merging this incomplete set of tests for several reasons:

* I believe that all the low-hanging fruit (such as the crash with
  varargs and a wrong type) has been fully picked.

* There are currently two issues that involve tests. I want each of
  them to be a different commit. Both of them involve areas not yet
  tested, so this would be incomplete anyway.

* This provides more tests to run Octeon up against.