Add a prototype struct to function values.
It holds the full name of a function as well as the full path of the
module. I've used the former to do some simplification around printing
the name of a function.

The idea, overall, is to use protos more in the future so that
creating a closure is less expensive.