Add a code pointer and locals to protos.
The first is to make tearing down code easier: Protos are 1:1 with
bytecode whereas function vals are N:1 when there are closures.

The second is to further reduce what closures need to copy.

I've undone the union around cid_table and upvalues because setting
the upvalue count to uint16_t -1 feels hacky. I tried getting rid of
zero upvalue closures, only to find that it's too difficult to be
worth the small gain.