Update the builtin package for the latest bindgen.
The latest bindgen adds generic information to foreign classes. This
provides generic information to `List` and `Hash`. However, those
classes are manually initialized internally (see lily_pkg_builtin's
bottom). This information is therefore entirely useless. The purpose
of this commit is to prevent the bindgen update from showing up in an
unrelated patch.