Private class members no longer pollute parent namespaces.
It used to be that declaring a private member of a base class would
prevent it from being declared in an inheriting class. That's bad,
because it means that class writers have to be careful about the names
they pick for -private- vars/methods.

This patch allows overriding private members in an inheriting class.
At the same time, it also provides a warning when trying to call a
private member from an upward class instead of allowing it to succeed.

Internally, this is done by making nested lookup of members no longer
return private members of base classes. Instead, the caller must do
another lookup to get the scope of the member (with the knowledge that
if the lookup succeeds, it's a private member). I've done it that way
in hopes of eventually removing class from properties and vars.