Proof of why lambdas returning the self type is bad. (#334).
Class methods take 'self' as their first parameter which is necessary
for the self return to work. Lambdas don't have the same guarantee.
This provides a proof explaining why allowing lambdas to do that is
bad.

This doesn't solve the above issue, but will make it easier to do so.