Dynaload of scoped enums, plus some dynaload coverage (fixes #322).
This adds 'covlib' which is to become a home for any coverage that
involves dynaloading. Currently, it exports a scoped and a flat enum
for testing.

I've also added another kind of test that runs from not within the
testing directory. The test kind can be removed once import can
understand `../` in paths.

This, to me, is enough testing of dynaload, as a previous patch fixed
and covered interpolation of a foreign class. That was the other
reason that the above issue was added.