It's time for lily_move to go.
lily_move is a very low-level way of placing a raw value into a whole
one. The only user of this is and should be the vm. All others should
use the api.

This takes the parts of lily_move that the vm is using and pulls them
into lily_vm.c. They're static now to prevent other parts of the
interpreter from using them.

There's room for simplifying what remains, but that'll be for another
patch.