Drop covlib tests until #336 lands.
Travis and AppVeyor both create a build directory, so the directories
are different. This causes the tests to work for me, but fail on the
two above platforms.

Once Lily understands relative imports, attaching `../` to the front
of the import paths should work. If it doesn't, then I'll think of
something else.