Emitter: Simplify call handling by dropping the first tree.
From the beginning, calls have always had their arg start as their
source. But later on, `x.y` calls and method calls from another
method made it complicated. It added a lot of complexity in call
running and writing.

Now it's gone. begin_call now cuts away the first argument if it's
just a source and not part of the run. It involves storing more info
on the source call tree (it's loaded with info now).

I'm doing this to make keyword arguments easier to merge in.