Simplify how self works in emitter to really fix #346.
The original fix included an invalid read because I didn't test it as
well as I should have. This goes well beyond that.

I've adjusted how 'self' works in emitter because the way it worked
before was broken and prone to issues. Now, the only 'self' that is
valid is the one that is drawn from a function block. Drawing it from
a non-function block is now wrong (no longer bubbles up).

The question of "do I need to draw self from a closure" is now as
simple as "does the current function block have a self?".