Keyword arguments round 2 (optargs and documentation). (fixes #332).
Before this patch, there was a barrier around trying to put these two
together. The reason is that optargs and keyargs present a case where
a function call may have missing parameters in between parameters that
are provided. The vm wasn't equipped to deal with that, so the feature
was turned off for the time being.

This patch adds a new class id: LILY_ID_UNSET. When emitter has to
make a call with argument holes, the holes are filled by taking a
register of a hidden unset type. At vm-time, foreign functions who
want to mix the two features will need to test both the argument
count, and for the arg being unset.

This patch also adds overall documentation for keyword arguments, now
that it works smoothly with other call features.