Forward definitions (fixes #337).
This adds the `forward` qualifier for `define`. A forward definition
is one that can be used immediately, but most be resolved before the
end of the scope in which it is defined. This is an essential part of
getting mutual recursion in Lily.

Currently, `import` and var declaration are blocked if there are
unresolved forwards. It's a simple strategy for making sure that a
forward define can't use a var that won't be initialized in time.

I've also added a section on modifiers to the tutorial, covering both
`forward` and other modifiers currently present.