String literal import paths (fixes #347).
Previously, it was possible to have slashed paths that were unquoted.
This patch is a breaking change, requiring those paths to now be
quoted. On the upside, string literal paths mean that relative paths
`../` now work. With that, some coverage tests for enum have been
uncommented and run with no modifications except for path changes.