Fix nested files not able to load a base package (fixes #338).
Prior to this, imports worked from a relative path. Importing from a
higher up directory could be done by adding `../` to the base of the
import path.

However, an issue was recently discovered with this approach:

```
packages
    mylib
        src
            mylib.lily

somedir
    abc.lily

library.lily
```

Suppose that `library.lily` imports `abc.lily`. The latter wants to
use `mylib.lily`. Under the old importing mechanism, that was
impossible. `abc.lily`'s imports ran relative, and adding `../`
doesn't work for packages.

This changes the import system to consider the beginning of any
package to be a root. Imports now run relative to a package root
directory.

An unfortunate side-effect is that `abc.lily` has to use `somedir/` as
a prefix for importing sibling modules. However, now `abc.lily` can
import `mylib.lily` which is far more important.

Sorry for doing this so close to a release.