Fix potential crash with optional foreign call (fixes #365).
The crash is very hard to trigger. You need a foreign function taking
optargs, and to pass a difference that's more than the number of spare
registers. So if there are 32 registers available and you pass 1 arg
out of a potential, say, 64, the interpreter crashes from trying to
clear out registers that foreign functions don't care about anyway.