vm: Do a gc drop after blasting the registers.
The vm used to do a gc walk before clearing out the registers. That
works for clearing values that have been lost. But it fails to drop
values that are circular and still within a register. A good example
of this is a class instance that has one property (a `Some` of
itself). It's still in a register and has a gc tag which caused the
old strategy to leak.

The new strategy is to blast the registers, then destroy any values
that are still alive. This new strategy is safe because the tagged
values hanging on don't have any attachment back to the destroyed
registers.

A test has been added that, using the previous strategy, caused a
memory leak.