ts: Take another shot at function's scoop collection.
Previously, function's collect would simply give a thumbs up if it saw
scoop on the left side. That's fine when a type has scoop on the left
but doesn't end up resolving or using it later at all. But a coming
patch resolves scoop.

This patch fixes function to collect any arguments missing on the
right side if scoop is the last missing argument on the left side. It
also adjusts the optarg checking logic.

It's not pretty, but it's necessary for future Coroutine work.
Hopefully I'll find a way to make it prettier later.