TYPE_HAS_OPTARGS now bubbles up from optarg types.
tm now only needs vararg information when potentially building a call
type. Optargs now bubbles up, so it doesn't matter if the calling type
wrote that down or not. I haven't adjusted parser's type collection
from using it, since it's still useful for said collection.

Parser is responsible for tagging optarg types, which is easy because
it's only done in one place. It seems reasonable to do that since the
optarg type is the only one which needs to carry over a flag to the
type based on its class.

This is necessary because scoop may collect optional arguments and
rebuilding optarg status on a type is harder than letting that status
bubble up.