Coroutines at last (fixes #273).
A coroutine is a function that can be suspended and executed later.
Internally, this works by having a sub vm wrapped up into a value and
providing a certain set of functions over it.

The type system needed to be adjusted to unroll scoop types, since
coroutines make use of scoop types to perform a transformation.

Most coroutine functions are stored in lily_vm.c, since they need to
use internal vm functions.

When a new state is created, it now holds a 'global' state. The
'global' state is handed down to any coroutine vm's of the state. This
allows coroutine vm's to access global vars, add to the gc, and
ensures that coroutines do not use stale tables. This adds a very small
(one pointer more) cost to a handful of areas in the interpreter.