Get rid of Coroutine.create while there's still time. (sorry)
I was writing the docs for coroutines when I realized how gross that
`Coroutine.create` is. What it does is to create an intermediate
builder so that coroutines can get any number of extra arguments.

Getting it to work involved nudging the type system's scoop
collection so that I could have a result that uses scoop again.

I've removed `Coroutine.create` in favor of `Coroutine.build` (no
one-time args) and `Coroutine.build_with_value` (a single one-time
arg). If a coroutine wants 2+ starting arguments, it can pass a single
Tuple for the caller to unpack or use a closure.

In addition to having to hack ts, `Coroutine.create` also made an
intermediate builder `Function`. So every coroutine made would have
a `Function` made as garbage _and_ a second gc tag. Let's not do that.

This commit reverts the ts work that `Coroutine.create` needed. I'm
open to adding the ts work back in if someone requests it and provides
a very good use case for it.