Add List.zip, plus two necessary emitter adjustments.
The new function makes use of Lily's scoop type to allow zipping any
number of `List` values. Since Lily doesn't have a nil/unset kind of
value, I've decided that the result will be the size of the smallest
`List` provided.

The first emitter adjustment is to bring back accounting for scoop
when handling a return type.

The second adjustment is to have empty `List` values check for
inference being the scoop type. That wasn't possible before, but now
it is by sending `[]` as an argument to `List.zip`. I've decided to
have such a case return `Unit` since that seems reasonable.