Lily's Random library license issue fixed (fixes #369).
The old library was licensed using BSD 3 clause. This collides with
Lily's license being MIT. I'd like for people to be able to embed or
use Lily however they wish without worrying about a component of Lily
using a license that requires attribution or effort that Lily itself
doesn't require.

I've picked a library that uses The Unilicense, which is effectively
public domain. One drawback of the change is that the new library only
allow for 32-bit ranges. It's thought that this won't be a big issue,
since most users of randomness don't need such a far reaching range.