Remove `Dynamic` from the language (fixes #371).
This class has existed in the interpreter since very early on. As time
has gone on, it's been harder to justify its existence. It's opt-in,
it requires a gc tag, and it can only store monomorphic types. Maybe
someone will make an argument for it coming back in the future.

For the time being, Lily doesn't have the facilities to make this
class useful, and such facilities are not what I want to put into the
language.