Dynaload now uses info and call tables (fixes #373).
Dynaload needed a loader function because a long time ago vars and
methods loaded a different kind of a value. But now vars are
implemented by calling a function that has the same signature as a
method. So now, bindgen creates a call table and the interpreter uses
that instead of a loader.

Existing libraries sources will need to have bindgen rerun on them.
The interpreter will now refuse to load old libraries.