api: Add lily_call_entry_func, fix last commit breaking the build.
Before this commit, the call table was defined as a void pointer in a
couple areas in the api. This lead to the compiler ignoring incorrect
definitions, resulting in the load embedder tests compiling but then
crashing hard.

All bindings have been redone with an additional patch that removes
the OFFSET definitions that were rendered useless when the loader was
taken out.