Goodbye, old Garden.
The old Garden was a good starting point for helping discover what I
really wanted from a package manager for this language. Since then,
the language has grown enough that I've taken the time to write a
more powerful replacement.

I've put a link to the new Garden in the README.