Values don't store class id in flags now (#327).
Prior to this patch, values in the vm stored class id in the lower 16
bits of their flags. The upper 16 bits of flags were reserved for
flags like refcountedness and gc marks.

With this patch, class ids have been replaced with a set of base
markers. Each value belongs to one of about 17 or so different groups,
and gets a corresponding value in that range. Some groups also have a
flag too. That's because some groups, like Integer, are useful to have
a quick membership test flag for.

Empty variants store their class id in the field that Integer uses.
Everything else that has a class id of interest should be a container,
and thus have a class_id field with it set.

This patch is rough in some places. Value pushing could be made
simpler, possibly. There's also potential for more kinds of flags.

Originally, the goal of this patch was to get rid of problems that
Lily was having when being used on a big endian system. However, it
turns out that this may have some performance potential in the long
run.