Build coverage library in test/ so scripts can use it.
The test that were using the coverage library were commented out
because they weren't working under CI. That's because CI builds Lily
in a different directory, and the library wasn't in a universally
relative spot from the pre commit harness.

Now that the coverage library builds in the testing directory, it can
be used by scripts once again. A couple tests that were written for it
have been uncommented. More tests using covlib will be coming soon.