emitter,testing: Better inference with calls based on result.
If the result of a call is unsolved and there's inference to use, try
it with ts check. If that fails, there may at least be more useful
information than not trying it at all. Guarding it with a type class
id check in emitter is a bad idea because ts is really good at
comparing type (emitter isn't).

Changed fail_variance to fail_type_check and added a test that
depends on this. One existing test needed to be changed because
better inference caused it to have a better (thus different) error
message.