symtab: Get rid of this weird implicit boxed lookup thing.
I checked the git history. This was part of the initial checkin for
implicit lookups. I think what this is trying to do is to make it
where referencing a flat enum will allow access to underlying parts. I
don't know why I would have added that.

Not only do I not want that behavior, it is currently not able to be
activated. Classes don't mark themselves with the right item kind.
That's okay though, because nothing cares.

Just get rid of it, I don't even like to look at it.