parser,testing: Don't stop import on invalid char (fixes #381).
This one is pretty bad. If an imported file had an invalid character
somewhere in it, the parsing of that file would stop and jump out of
the file. This is because, internally, that wasn't being checked for.

I found this by trying to add coverage to test/verify_string.lily. I
don't know how it happened, but a backtick snuck into the very start
of the file, causing the entire following set of tests to never be
added.

A test has been added to make sure this doesn't occur again. Yikes.