builtin: Fix File.read_line (break) (fixes #382).
The old version of `File.each_line` did not include the newline in
the lines it sent back. This made it inconsistent with how most
languages work and with `File.read_line`.

The former version of `File.each_line` also neglected to send the line
on which the final EOF was found.

This breaks scripts the rely on `File.read_line` which was previously
untested.