syntax: Constructor vars now require a scope (break) (fixes #400),
The previous version made it optional to specify a scope for a
shorthand constructor var. The last version did not require it
because the idea came late into the release (and it was an even
release too).

Code that relied on automatic promotion to public is now broken. This
change is being done so that variable declaration is consistent in
requiring a scope throughout the language.