Errors: Better traceback for foreign libraries (fixes #399).
Prior to this, the 'path' for foreign and registered modules was
`[C]`. That's okay when the source is obvious like a builtin class
method with the class name attached. However, it's not as helpful when
there's no classname (such as a `math.fmod`).

With this change:

* Registered modules now write their name in brackets.

* Libraries write the path they were loaded from.

The latter is an adjustment to the original issue message in the
interest of more clarity.