internals: Line numbers are unsigned 16, not 32.
65K should be enough for everyone. I bumped this to 32 a long time ago
because I thought that would be better for traceback. But really,
files of that size should be discouraged since this is a high level
interpreted language.

Additionally, the bytecode is unsigned 16 bit ints. So even if the
frontend supported 32 bit line numbers, the bytecode would have line
numbers wrapped around after 65535.

This patch does not do any cleaning up so that this is an isolated
change.