parser: Use the proper path in module lookup (fixes #423).
The interpreter creates paths with dot-slash at the front to prevent some
platforms from trying system library paths. However, all path storage is of
the raw path since the dot-slash is useless.

The above bug was caused by the dot-slash path being used for lookup. Since
the path was different, a new module entry was created instead of using the
old one. The end result was an incorrect type error.