api: Drop copy_str_input (fixes #427).
Previously, the copy_str_input option in config to determine if string
content passed to the interpreter should be copied.

The new approach is to not copy string input passed to render or
parse functions. This is because the string data passed to them is
likely to survive until processing is done because it's likely to be
on the stack somewhere.

On the other hand, string input passed to import functions is always
copied. This is because in cases like spawni, the string data may be
a non-literal string that might be wiped. Granted, that would be a
corner case, but it's still possible.

This new approach is, I think, more straightforward and more obviously
correct. It also satisfies the use cases mentioned in the commit that
first brought this config option.