Syntax: Static methods can now use class methods (fixes #430).
Static methods should be able to call class methods if they can
provide a suitable value for 'self'. Prior to this, attempts to call
non-static from static would complain about not having a self to
capture over.

For reference, Python allows the same behavior so I wouldn't call it
too out of the norm.

One test was moved from a fail to a feature (with some adjustments).