internals: Import funcs account for missing 'use' (fixes #431).
The api docs used to simply suggest always calling an import directory
selection function before calling an import. They never said why or
what the consequences of failing to do so were.

The api now clearly documents that failing to call a directory
selection function (one of the `lily_import_use_*` functions) will
cause the `lily_import_*` functions to be a no-op (except for the one
that loads library data).