symtab: item_kind marking for classes and enums.
Previously, only a handful of predefined classes had an item_kind
properly set to ITEM_TYPE_CLASS. This extends that to all classes. It
also adds ITEM_TYPE_ENUM and sets that on all enums.

This change was done in part for consistent internals, but the driving
reason is to allow introspect to easily walk boxed symbols.