api: Pushed containers are now gc speculative (fixes #433).
It's possible that a container created by the push api will be the
sole source of a gc tagged value. Previously, the tagged values would
be missed, causing them to be incorrectly destroyed.

This doesn't include testing because there aren't metrics to make sure
that said testing is actually working.

Though this modifies the pushing api functions, callers should not
notice a differece (aside from not crashing in rare cases, lol).