pkg_builtin: Make follower_table uint8_t, remove negatives (#327).
Previously, the follower table was an array of char that used -1
values to indicate an invalid state. I have a feeling this is causing
issues with less-common architectures. The table now uses '0' for the
invalid state and is defined as an array of uint8_t.

Lily expects that String values are proper utf-8, so this should not
affect the other String functions that use the follower_table.