vm: All traceback now uses 'in' instead of 'from' (fixes #419).
There are two areas where traceback is constructed: One in parser, and
the other in the vm. The two had a subtle format divergence where
parser was saying '<x> in <y>' and vm was saying '<x> from <y>'.

I noticed this when working on testkit since testkit needs to show the
traceback of unexpected errors. Testkit is currently doing a string
replace, but eventually I'd like it to use the traceback directly.

It's the small details that matter sometimes. :)