core_types: Add line_num for classes, enums, and variants.
The uint16_t line_num records where the class, enum, or variant was
declared. The field is currently unused by the interpreter, but a
future version of the interpreter can use it for better error
messages.

The line number is set similar to how it is for vars: Native code uses
the line number at the time of declaration, whereas dynaloads use 0.