vm: Fix potential invalid read with lily_push_super.
This invalid read is tough to come across and verify, so no test has
been added. It's caused by a test that creates an Exception through
lily_push_super. The same test creates a custom class that inherits
Exception that wants to use the same temp slot. So there's an invalid
read because the 'create a new instance' path for lily_push_super was
not initializing the ctor remaining to zero (since regular instances
do not need that).