cmake: Add Findlily.cmake for use by extension libraries.
It's currently difficult to write an extension library that works on
Linux and Windows. Many extension libraries are written with Linux in
mind and therefore fail to load on Windows.

This file provides a cross-platform means of finding the interpreter's
information. It also provides `lily_add_library` which provides a
cross-platform way of creating a library and also sets common options
that extension libraries set.