testing: HUGE rewrite of the testing suite (fixes #444).
This is a massive commit that completely changes how tests are
organized and executed. The focus of this commit is on introducing the
new test system, so most of the old tests have been migrated over.

The new test suite is driven by C. The driver has a hardcoded list of
directories and targets it expects to find. Each target imports a
testing file that defines a customizable subinterpreter and a TestCase
for running tests. Targets define a set of public methods, but don't
execute any code. Execution is done by the driver writing code to
invoke the target's test class.

It's a design that's similar to testkit. It makes use of testkit's
code as well as spawni's code, but simplifies both so that there are
no external dependencies.

Right now, the main advantages that the new suite brings are that it's
more organized and it can simulate imports. It's a solid framework
that's necessary to provide deeper testing needed by an ever-growing
interpreter.