interp: Organize exception internals to fix a bug (fixes #439).
The bug in question is the message returned when asking an interpreter
for the error message after a successful parse. Prior to this, it
would read similar to this:

```
Error
    from [test]:0:
```

The interpreter now returns an empty string if there is no error
message.

Internally, this involved cleaning up how the interpreter uses the
raiser:

* The raiser now holds an enum to signal where the issue came from.

* The vm now only uses the raiser if exception handling fails.

* Emitter sends the failed tree instead of setting a line adjust.

* Parser's error building has been cleaned up.