parser: Better class/enum/variant redeclaration message (fixes #452).
If a class is declared with the same name as an enum, the interpreter
now mentions an enum was the prior source (versus always saying a
class).

Redeclaration messages now also mention if the source was a predefined
class/enum/variant. This fixes a small nit wherein using a predefined
class like `Exception`, then attempting to declare it again had a
different error message than directly attempting a redeclaration.

This was discovered while working on #389 (better rewind testing).