internal: The class for `Unit` is now in the symtab.
The `Unit` class needs to be easily accessible, so it's global and
readonly. To make it work with the symtab, it's attached at the end of
the first class (`Integer`). When the symtab is being torn down, it's
unlinked from the symtab to keep it from being dropped.

With this, `Unit` is now reachable like any other class, removing
possible inconsistencies with different areas doing different
searches.