internal: Make identifier lookup consistent (fixes #396).
This is yet another internal cleanup which has been a long time
coming. This time the target is how parser and symtab work together to
discover symbols. In cleaning up this area of the interpreter, a few
issues were discovered that were fixed as part of this commit.

First and most importantly, searches are now done only in the module
given. Prior to this commit, sending a NULL module would search the
builtin module as well as the active module. This more explicit
behavior has helped to make expression and other search-heavy parts of
the interpreter more clear. Searches are only done in the modules that
are given, no magic behind-the-scenes behavior.

Speaking of which, symtab no longer takes the generic pool since it no
longer does generic class lookups. That is now parser's concern,
especially since there are limited areas that need to search for a
generic class.

Member redeclaration and private member checks have been unified to
make them more consistent. Member lookup has also been expanded upon
in the areas where it's done. Shadowing detection is better now, which
may cause some code to fail where once worked. However, fixing such
code is trivial -- simply rename the offending method or property.

Plenty of tests have been added to cover the new paths.