emitter: List/Hash inference adjustment (fixes #410).
With this patch, inference is ignored if the first element does not
agree. In such situations, there's almost certainly going to be an
error somewhere. However, it's incorrect to say that List/Hash literal
values are the mismatching spot.

The result is better error messages, since the error should occur at
the point which sent the inference down.