parser: Allow redeclaration of super privates (#396).
When I pushed #396, I thought the correct behavior was to warn when
members were redeclared, full stop. So I adjusted the interpreter's
testing suite accordingly as part of that patch. But it turns out that
this is the correct behavior. It was being done before, but not
applied in all member declaration contexts.

The only exception is shorthand, because shorthand properties come
before inheritance, so having a visibility check is testing for the
impossible.

This was discovered by trying to run testkit. New tests have been
added to make sure this doesn't slip by again.