tooling: The prelude is now autogenerated and outside parser.
Prior to this, parser needed to extern in the data tables of prelude
modules in order to call module registry functions on them. This adds
a script that autogenerates a prelude init function for parser to
call.

Since this is only one of the steps involved in adding a new module to
the interpreter's prelude, the script also includes a checklist in the
comments. This should make it easier to add new modules to the
interpreter's prelude.