parser: Merge parse_class_body into keyword_class, plus cleanup.
A long time ago, dynaload initialized `Exception` by writing the text
for it into a buffer and calling for a parse. Because of that, class
parse was split into two functions. Since that code is gone and will
never come back, this merges the functions and removes now-unnecessary
save/restore code.